changes float/double

+++ src/SideEuler.cu
@@ -106,8 +106,8 @@ __host__ void EvanescentBoundary (float *Vrad, float *Vtheta, float *Dens, float
 
   float Tin, Tout, DRMIN, DRMAX;
   /* Orbital period at inner and outer boundary */
-  Tin = 2.0*PI*powf(Rmed[0],3./2);;
-  Tout = 2.0*PI*powf(Rmed[NRAD-1],3./2);
+  Tin = 2.0*PI*pow(Rmed[0],3./2);;
+  Tout = 2.0*PI*pow(Rmed[NRAD-1],3./2);
   /* DRMIN AND DRMAX are global Radii boundaries of killing wave zones */
   DRMIN = Rmed[0]*1.25;
   DRMAX = Rmed[NRAD-1]*0.84;
